+++ CHANGES.rst
@@ -19,6 +19,17 @@ Release History
    - Bugfixes
    - Documentation
 
+3.0.0 (unreleased)
+==================
+
+**API changes**
+
+- Learning rules take ``pre_synapse``, ``post_synapse``, and ``theta_synapse``
+  instead of ``pre_tau``, ``post_tau``, and ``theta_tau`` respectively.
+  This allows arbitrary ``Synapse`` objects to be used as filters on the
+  learning signals. The ``learning_rate`` parameter also always comes first.
+  (`#1095 <https://github.com/nengo/nengo/pull/1095>`_)
+
 2.2.0 (unreleased)
 ==================